Introduce check_ceph_libs check
Introduce check_ceph_libs_mk, a local checkmk check which looks for
processes who have ceph-related shared libraries mapped that no longer
exist.

This condition can happen when upgrading ceph-related packages on a
client node and not restarting affected processes. This can cause
serious problems to the client, such as inability to talk to Ceph, stuck
I/O, archipelago issues, etc.

The check is written in Python and depends on psutil library, which is
already installed on our fleet. Using a list of predefined process names
and libraries, it iterates over the complete process list of the node,
filter out any unrelated processes and then, for each process, looks up
its memory maps for '(deleted)' entries.

Also, this script has a 'hidden' argument: '--show'. Calling it with it,
an operator can see a complete list of processes using stale libraries
in JSON.