check_ceph_libs_mk: Work around (another) regression caused by psutil
Since v4.0.0 [1], psutil strips the ' (deleted)' string from mapped
libraries that have been deleted from the disk. That renders our check
useless, because psutil has no way to tell about this info.

Work around this behavior by reading the maps file of each target
process in procfs. We decided not to run subprocesses and parse output
from linux utilities, but the behavior of this library forces us to do
so.

[1] https://github.com/giampaolo/psutil/issues/759