check_ceph_libs_mk: Ignore QEMU procs with Archipelago disks
QEMU contains native support for RBD and its enabled by default on
Debian systems. That means that librbd and librados are loaded
automatically, even if we do not use them.

Right now, our check looks up if QEMU processes with RBD or
Archipelago disks have stale libraries. With Archipelago disks, this
is useless, because QEMU does not use librados/librbd in any way; it
communicates Archipelago using tapdisk/blktap. This can cause false
positives to pop up.

So, modify check_ceph_libs so it does not check QEMU processes with
Archpelago disks.

Kudos to philipgian for all the info.